FederationをPage側で状態保持するように修正
