+++ lib/view/federation_page/federation_info.dart
@@ -2,14 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:miria/extensions/string_extensions.dart';
 import 'package:miria/model/federation_data.dart';
-import 'package:miria/providers.dart';
-import 'package:miria/view/common/account_scope.dart';
 import 'package:miria/view/common/constants.dart';
 import 'package:miria/view/common/error_detail.dart';
 import 'package:miria/view/common/misskey_notes/network_image.dart';
 import 'package:miria/view/federation_page/federation_page.dart';
 import 'package:miria/view/themes/app_theme.dart';
-import 'package:misskey_dart/misskey_dart.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class FederationInfo extends ConsumerStatefulWidget {
@@ -27,123 +24,6 @@ class FederationInfo extends ConsumerStatefulWidget {
 class FederationInfoState extends ConsumerState<FederationInfo> {
   (Object?, StackTrace)? error;
   FederationData? data;
-
-  @override
-  void didChangeDependencies() {
-    super.didChangeDependencies();
-    Future(() async {
-      try {
-        final account = AccountScope.of(context);
-        if (widget.host == account.host) {
-          // 自分のサーバーの場合
-          final metaResponse = await ref.read(misskeyProvider(account)).meta();
-          final statsResponse =
-              await ref.read(misskeyProvider(account)).stats();
-          ref.read(federationPageFederationDataProvider.notifier).state =
-              FederationData(
-                  bannerUrl: metaResponse.bannerUrl?.toString(),
-                  faviconUrl: metaResponse.iconUrl?.toString(),
-                  tosUrl: metaResponse.tosUrl?.toString(),
-                  privacyPolicyUrl: (metaResponse.privacyPolicyUrl)?.toString(),
-                  impressumUrl: (metaResponse.impressumUrl)?.toString(),
-                  repositoryUrl: (metaResponse.repositoryUrl).toString(),
-                  name: metaResponse.name ?? "",
-                  description: metaResponse.description ?? "",
-                  usersCount: statsResponse.originalUsersCount,
-                  notesCount: statsResponse.originalNotesCount,
-                  maintainerName: metaResponse.maintainerName,
-                  maintainerEmail: metaResponse.maintainerEmail,
-                  serverRules: metaResponse.serverRules,
-                  reactionCount: statsResponse.reactionsCount,
-                  softwareName: "misskey",
-                  softwareVersion: metaResponse.version,
-                  languages: metaResponse.langs,
-                  ads: metaResponse.ads,
-                  meta: metaResponse,
-
-                  // 自分のサーバーが非対応ということはない
-                  isSupportedAnnouncement: true,
-                  isSupportedEmoji: true,
-                  isSupportedLocalTimeline: true);
-        } else {
-          final federation = await ref
-              .read(misskeyProvider(AccountScope.of(context)))
-              .federation
-              .showInstance(FederationShowInstanceRequest(host: widget.host));
-          MetaResponse? misskeyMeta;
-
-          bool isSupportedEmoji = false;
-          bool isSupportedAnnouncement = false;
-          bool isSupportedLocalTimeline = false;
-
-          if (federation.softwareName == "fedibird" ||
-              federation.softwareName == "mastodon") {
-            // already known unsupported software.
-          } else {
-            try {
-              // Misskeyサーバーかもしれなかったら追加の情報を取得
-
-              final misskeyServer =
-                  ref.read(misskeyWithoutAccountProvider(widget.host));
-              final endpoints = await misskeyServer.endpoints();
-
-              if (endpoints.contains("announcement")) {
-                isSupportedAnnouncement = true;
-              }
-
-              // 絵文字が取得できなければローカルタイムラインを含め非対応
-              if (endpoints.contains("emojis")) {
-                isSupportedEmoji = true;
-
-                if (endpoints.contains("notes/local-timeline")) {
-                  isSupportedLocalTimeline = true;
-                }
-              }
-
-              misskeyMeta = await misskeyServer.meta();
-            } catch (e) {}
-            ;
-          }
-
-          ref.read(federationPageFederationDataProvider.notifier).state =
-              FederationData(
-            bannerUrl: (misskeyMeta?.bannerUrl)?.toString(),
-            faviconUrl: (federation.faviconUrl)?.toString(),
-            tosUrl: (misskeyMeta?.tosUrl)?.toString(),
-            privacyPolicyUrl: (misskeyMeta?.privacyPolicyUrl)?.toString(),
-            impressumUrl: (misskeyMeta?.impressumUrl)?.toString(),
-            repositoryUrl: (misskeyMeta?.repositoryUrl)?.toString(),
-            name: misskeyMeta?.name ?? federation.name,
-            description: misskeyMeta?.description ?? federation.description,
-            maintainerName: misskeyMeta?.maintainerName,
-            maintainerEmail: misskeyMeta?.maintainerEmail,
-            usersCount: federation.usersCount,
-            notesCount: federation.notesCount,
-            softwareName: federation.softwareName ?? "",
-            softwareVersion:
-                misskeyMeta?.version ?? federation.softwareVersion ?? "",
-            languages: misskeyMeta?.langs ?? [],
-            ads: misskeyMeta?.ads ?? [],
-            serverRules: misskeyMeta?.serverRules ?? [],
-            isSupportedEmoji: isSupportedEmoji,
-            isSupportedLocalTimeline: isSupportedLocalTimeline,
-            isSupportedAnnouncement: isSupportedAnnouncement,
-            meta: misskeyMeta,
-          );
-        }
-
-        if (!mounted) return;
-        setState(() {});
-      } catch (e, s) {
-        print(e);
-        print(s);
-        if (!mounted) return;
-        setState(() {
-          error = (e, s);
-        });
-      }
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
